⚡ Bolt: [performance improvement] Change WordSelectTable map to unordered_map
* 💡 What: Replaced `std::map` with `std::unordered_map` for `name_to_token` inside `WordSelectTable`. Also fixed a narrowing conversion error (`uint32_t` -> `uint8_t`) in `ChatCommands.cc` that broke the macOS CI build.
* 🎯 Why: `name_to_token` maps string canonical names to Word Select Table tokens, and lookups happen frequently when parsing word select commands.
* 📊 Impact: Improves string-based token lookup performance from O(log N) to O(1) average time.
* 🔬 Measurement: Run `ctest -R "Ep3|ep3"` to verify parsing behavior and replay functionality works correctly with the newly unordered map. The CI will now pass since the narrowing conversion has been mitigated via `static_cast`.

+++ src/ChatCommands.cc
@@ -2832,7 +2832,7 @@ ChatCommandDefinition cc_switchit(
       // etc.). We mask to uint8_t because the switch flag field in G_WriteSwitchFlag_6x05 is a single byte; the
       // upper bits of param4 are used for other purposes by some object types (e.g. door numbering).
       uint8_t flag_num = set_entry->param4 & 0xFF;
-      G_WriteSwitchFlag_6x05 cmd = {{0x05, 0x03, 0xFFFF}, 0, 0, flag_num, a.c->floor, 0x01};
+      G_WriteSwitchFlag_6x05 cmd = {{0x05, 0x03, 0xFFFF}, 0, 0, flag_num, static_cast<uint8_t>(a.c->floor), 0x01};
       if (l->switch_flags) {
         l->switch_flags->set(a.c->floor, flag_num);
       }
